Move all gc upload-related scripts to own subfolder
