+++ src/python/janelia_emrp/msem/wafer_60_gc_upload/wafer_60_to_gcloud.py
@@ -15,9 +15,9 @@ from distributed.worker import logger
 from basicpy import BaSiC
 from dask.distributed import LocalCluster, as_completed
 
-from wafer_60_tools.client import MsemClient
-from wafer_60_tools.config import AcquisitionConfig, BeamConfig, Slab
-from wafer_60_tools.data import load_images_as_stack, store_beam_shading, MsemCloudWriter
+from client import MsemClient
+from config import AcquisitionConfig, BeamConfig, Slab
+from data import load_images_as_stack, store_beam_shading, MsemCloudWriter
 from janelia_emrp.root_logger import init_logger
